+++ src/clone_node.ts
@@ -1,8 +1,15 @@
-import { Maybe, maybeOf } from '@fluss/core';
+import { isMaybe, Maybe } from '@fluss/core';
 
+export function cloneNode<T extends Node>(
+  node: Maybe<T>,
+  deep?: boolean
+): Maybe<T>;
+export function cloneNode<T extends Node>(node: T, deep?: boolean): T;
 export function cloneNode<T extends Node>(
   node: T | Maybe<T>,
   deep: boolean = false
-): Maybe<T> {
-  return maybeOf(node).map((node) => node.cloneNode(deep) as T);
+): T | Maybe<T> {
+  return isMaybe<T>(node)
+    ? node.map((node) => node.cloneNode(deep) as T)
+    : (node.cloneNode(deep) as T);
 }