+++ test/add_event_listener.test.ts
@@ -43,9 +43,7 @@ describe('addEventListener', () => {
     querySelector<HTMLDivElement>('.second').map((element) => {
       addEventListener(element, 'animationcancel', (evt) => {
         event = evt;
-      }).map((removeFullScreenChangeListener) =>
-        removeFullScreenChangeListener()
-      );
+      })();
 
       element.dispatchEvent(new Event('fullscreenchange'));